fix: изменен тип данных для id всех entity и dto на Long, добавлены экземпляры классов User и Event вместо id в класс Comment, в файле schema.sql изменен способ объявления primary key(кроме составных), изменен способ объявления правила unique для полей.

+++ ewm-service/src/main/java/ru/practicum/explore_with_me/request/RequestPrivateController.java
@@ -18,18 +18,18 @@ public class RequestPrivateController {
     private final RequestService service;
 
     @PostMapping("/{userId}/requests")
-    public ParticipationRequestDto saveNewEvent(@PathVariable long userId,
-                                                @RequestParam(name = "eventId") Integer eventId) {
+    public ParticipationRequestDto saveNewEvent(@PathVariable Long userId,
+                                                @RequestParam(name = "eventId") Long eventId) {
         return service.saveNewRequest(userId, eventId);
     }
 
     @GetMapping("/{userId}/requests")
-    public List<ParticipationRequestDto> getRequestByUserId(@PathVariable long userId) {
+    public List<ParticipationRequestDto> getRequestByUserId(@PathVariable Long userId) {
         return service.getRequestByUserId(userId);
     }
 
     @PatchMapping("/{userId}/requests/{requestId}/cancel")
-    public ParticipationRequestDto canceledRequestByRequester(@PathVariable long userId, @PathVariable long requestId) {
+    public ParticipationRequestDto canceledRequestByRequester(@PathVariable Long userId, @PathVariable Long requestId) {
         return service.canceledRequestByRequester(userId, requestId);
     }
 }
